requirements.txt
torch
torchvision
numpy
pandas
opencv-python
Pillow
streamlit
tqdm
opencv-python-headless
matplotlib
gdown